+++ docs/changelog.md
@@ -2,11 +2,11 @@
 
 ## Week of 29 July
 ### District Context App
-
+* Legislator lookup sorted by last name (not first name)
 
 
 ### Voting Patterns App
-* Add sort legislators by district electorate lean
+* Added sort legislators by district electorate lean
 
 
 ### ETL Pipeline